--- civis-csv/pom.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>com.civis</groupId>
-        <artifactId>civis-utils</artifactId>
-        <version>1.3.1-SNAPSHOT</version>
-    </parent>
-
-    <groupId>com.civis.civis-utils</groupId>
-    <artifactId>civis-csv</artifactId>
-    <version>1.3.1-SNAPSHOT</version>
-    <packaging>jar</packaging>
-
-</project>
