--- civis-csv/src/main/resources/log4j.properties
@@ -1,5 +1,0 @@
-log4j.rootLogger=DEBUG, STDOUT
-log4j.logger.deng=DEBUG
-log4j.appender.STDOUT=org.apache.log4j.ConsoleAppender
-log4j.appender.STDOUT.layout=org.apache.log4j.PatternLayout
-log4j.appender.STDOUT.layout.ConversionPattern=%5p [%t] (%F:%L) - %m%n
