+++ civis-util/pom.xml
@@ -19,7 +19,6 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-
     <parent>
         <groupId>com.civis</groupId>
         <artifactId>civis-tools</artifactId>
@@ -28,21 +27,8 @@
     </parent>
 
     <groupId>com.civis.civis-tools</groupId>
-    <artifactId>civis-opennlp</artifactId>
+    <artifactId>civis-util</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.apache.opennlp</groupId>
-            <artifactId>opennlp-tools</artifactId>
-            <version>1.5.3</version>
-        </dependency>
-        <dependency>
-            <groupId>com.civis.civis-tools</groupId>
-            <artifactId>civis-util</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
 
 </project>
