+++ civis-util/src/main/java/com/civis/utils/csv/common/CSVData.java
