+++ pom.xml
@@ -1,12 +1,28 @@
 <?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright 2015 Sergej Meister
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~  http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~  Unless required by applicable law or agreed to in writing, software
+  ~  distributed under the License is distributed on an "AS IS" BASIS,
+  ~  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~  See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.civis</groupId>
-    <artifactId>civis-utils</artifactId>
-    <version>1.3.1-SNAPSHOT</version>
+    <artifactId>civis-tools</artifactId>
+    <version>0.0.1-SNAPSHOT</version>
 
     <!-- Output to pom format -->
     <packaging>pom</packaging>
@@ -80,7 +96,7 @@
         </plugins>
     </build>
     <modules>
-       <module>civis-csv</module>
+       <module>civis-util</module>
        <module>civis-opennlp</module>
     </modules>
 </project>
